Day 215

+++ Game/Assets/Misc/Interfaces/IDamageable.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void TakeHealth(float intake, float shieldIntake);
-    float TakeDamage(float intake, Stats sender, ref float shieldOut, float arbParam, int arbParam2);
+    void TakeHealth(float intake, int shieldIntake);
+    float TakeDamage(float intake, Stats sender, ref int shieldOut, float arbParam, int arbParam2);
     float TakeDamage(float intake, Stats sender, float arbParam, int arbParam2);
     void TakeInjector(Injector injector, bool cacheInstances);
     void RevertInjector(Injector refInjector);
